UI 분리: ProfileScreen과 MoreScreen 기능 구조 정리
