+++ eatsoon/src/firebase/index.js
@@ -22,4 +22,4 @@ const app = initializeApp(firebaseConfig);
 const auth = getAuth(app);
 const db = getFirestore(app); // Firestore 추가
 
-export { auth, db };
+export { auth , db };
