+++ eatsoon/src/navigations/SettingsStack.js
@@ -2,7 +2,7 @@
 
 import React from 'react';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import SettingsScreen from '../screens/SettingsScreen';
+import MoreScreen from '../screens/MoreScreen';
 import RecipeRecommendationScreen from '../screens/RecipeRecommendationScreen';
 import StatisticsReportScreen from '../screens/StatisticsReportScreen';
 import ShoppingListScreen from '../screens/ShoppingListScreen';
@@ -13,7 +13,7 @@ export default function SettingsStack() {
     console.log('✅ SettingsStack loaded'); 
     return (
     <Stack.Navigator>
-      <Stack.Screen name="Settings" component={SettingsScreen} />
+      <Stack.Screen name="More" component={MoreScreen} />
       <Stack.Screen name="RecipeRecommendation" component={RecipeRecommendationScreen} />
       <Stack.Screen name="StatisticsReport" component={StatisticsReportScreen} />
       <Stack.Screen name="ShoppingList" component={ShoppingListScreen} />
