+++ eatsoon/src/screens/RecipeRecommendationScreen.js
@@ -1,22 +1,28 @@
 import React from 'react';
-import { View, Text, StyleSheet } from 'react-native';
+import { View, Text, Button } from 'react-native';
+import { db } from '../firebase';
+import { collection, addDoc } from 'firebase/firestore';
 
-export default function RecipeRecommendationScreen() {
+const RecipeRecommendationScreen = () => {
+  const addTestRecipe = async () => {
+    try {
+      await addDoc(collection(db, "recipes"), {
+        name: "된장찌개",
+        ingredients: ["된장", "두부", "애호박"],
+        createdAt: new Date()
+      });
+      console.log("레시피 저장 완료");
+    } catch (error) {
+      console.error("레시피 저장 실패: ", error);
+    }
+  };
+
   return (
-    <View style={styles.container}>
-      <Text style={styles.title}>레시피 추천 기능 준비 중</Text>
+    <View style={{ padding: 20 }}>
+      <Text>Recipe Recommendation Screen</Text>
+      <Button title="테스트 레시피 추가" onPress={addTestRecipe} />
     </View>
   );
-}
+};
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  title: {
-    fontSize: 20,
-    fontWeight: 'bold',
-  },
-});
+export default RecipeRecommendationScreen;
